requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.5.0
pytest==7.2.1
google-cloud-logging==3.2.4
